Move macro's in parser to common
This reduces the clutter, especially since documentation now is now
being added.
