+++ src/parser/call.rs
@@ -3,12 +3,12 @@ use crate::lexer::token::TokenPos;
 use crate::parser::_type::parse_id;
 use crate::parser::ast::ASTNode;
 use crate::parser::ast::ASTNodePos;
-use crate::parser::end_pos;
+use crate::parser::common::end_pos;
 use crate::parser::expression::is_start_expression_exclude_unary;
 use crate::parser::expression::parse_expression;
 use crate::parser::parse_result::ParseErr::*;
 use crate::parser::parse_result::ParseResult;
-use crate::parser::start_pos;
+use crate::parser::common::start_pos;
 use crate::parser::TPIterator;
 use crate::parser::_type::parse_id_maybe_type;
 use crate::parser::expression::is_start_expression;
