+++ src/parser/expr_or_stmt.rs
@@ -8,7 +8,7 @@ use crate::parser::control_flow_expr::parse_match_cases;
 use crate::parser::expression::parse_expression;
 use crate::parser::parse_result::ParseErr::*;
 use crate::parser::parse_result::ParseResult;
-use crate::parser::start_pos;
+use crate::parser::common::start_pos;
 use crate::parser::statement::parse_statement;
 use crate::parser::TPIterator;
